refactor(card): delete card and make card2 take its place
BREAKING CHANGE: delete card and make card2 take its place
