+++ lib/components/Card/Card.test.tsx
@@ -1,17 +1,17 @@
 import React from "react";
 import { render } from "../../utils/test-utils";
-import Card2 from "./Card";
+import Card from "./Card";
 
 describe("Card", () => {
   it("renders with default props", () => {
-    const { container } = render(<Card2 />);
+    const { container } = render(<Card />);
 
     expect(container.firstChild).toMatchSnapshot();
   });
 
   it("renders with custom props", () => {
     const { container } = render(
-      <Card2 padding="huge" shadow="upXxLarge" radius="xxhuge" />
+      <Card padding="huge" shadow="upXxLarge" radius="xxhuge" />
     );
 
     expect(container.firstChild).toMatchSnapshot();
