--- lib/components/Card/Card.types.ts
@@ -1,6 +1,0 @@
-/**
- * @deprecated in favor of Card2Props
- */
-export interface CardProps {
-  noPadding?: boolean;
-}
