+++ lib/components/Card/index.ts
@@ -1,9 +1,2 @@
-/**
- * @deprecated in favor of Card2
- */
+export type { CardProps } from "./Card";
 export { default } from "./Card";
-
-/**
- * @deprecated in favor of Card2
- */
-export type { CardProps } from "./Card.types";
