--- lib/components/Card2/index.ts
@@ -1,2 +1,0 @@
-export type { CardProps } from "./Card";
-export { default } from "./Card";
