+++ src/stories/components/table/Table.stories.js
@@ -56,7 +56,7 @@ export const SimpleTable = () => (
 );
 
 export const SimpleTableInCard = () => (
-  <Card noPadding>
+  <Card>
     <Table>
       <TableHead>
         <TableRow>
